refactor: merge comments-panel e2e specs, drop stale aria-expanded
- merge comments-panel-detail + comments-panel-redesign into a single
  comments-panel.spec.ts with two describe blocks; drop duplicate
  "Show resolved filter" test (covered by segmented filter test)
- review_live.html.heex: remove static aria-expanded="false" on
  #comment-count; JS owns it via syncCommentsPanelAria
